+++ packages/moxxmpp/lib/src/util/list.dart
@@ -1,0 +1,5 @@
+extension ListItemCountExtension<T> on List<T> {
+  int count(bool Function(T) matches) {
+    return where(matches).length;
+  }
+}