+++ packages/moxxmpp/lib/src/xeps/xep_0030/cache.dart
@@ -6,6 +6,7 @@ class DiscoCacheKey {
   const DiscoCacheKey(this.jid, this.node);
 
   /// The JID we're requesting disco data from.
+  // TODO(Unknown): Replace with JID
   final String jid;
 
   /// Optionally the node we are requesting from.